Split the UI initialization and fix the about box
Summary:
The KAboutData must be initialized and assigned before creating
the UI, otherwise the default about dialog in the Help menu shows
a default dialog.
Change the code to follow the usual structure: initialize the
Q(Core)Application, set the translation domain, set KAboutData,

An alternative fix could be changing KopeteApplication to
QObject and initializing Q(Core)Application separately, as it
was done in other programs; the current change aims to be less
invasive.

Test Plan:
The "About Kopete" menu entry shows all the information of
the application.

Reviewers: pali, mlaurent

Reviewed By: mlaurent

Subscribers: #kopete

Tags: #kopete

Differential Revision: https://phabricator.kde.org/D8319